refactor new project form flow
